增加lob失败时meta的toString打印
git-svn-id: http://192.168.20.11/sequoiadb/trunk@23478 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/include/dmsLobDef.hpp
@@ -200,8 +200,8 @@ namespace engine
          stringstream ss ;
          ss << "Len:" << _lobLen
             << ", CreateTime:" << _createTime
-            << ", Status:" << _status
-            << ", Version:" << _version
+            << ", Status:" << (UINT32)_status
+            << ", Version:" << (UINT32)_version
             << endl ;
          return ss.str() ;
       }
